+++ src/app/controllers/StudentController.js
@@ -114,9 +114,7 @@ class StudentController {
 
   async delete(req, res) {
     const schema = Yup.object().shape({
-      id: Yup.number()
-        .positive()
-        .min(1),
+      id: Yup.number().required(),
     });
     const { id } = req.params;
     // validate schema and return all errors messages if need
@@ -124,6 +122,12 @@ class StudentController {
       await schema.validate({ id }, { abortEarly: false });
     } catch (error) {
       return res.status(400).json({ errors: error.errors });
+    }
+
+    const student = await Student.findByPk(id);
+
+    if (!student) {
+      return res.status(400).json({ error: 'Student does not exists' });
     }
 
     await Student.destroy({ where: { id } });
